+++ core/src/routes/(modules)/options/text.svelte
@@ -44,6 +44,11 @@
 	import BorderRadius from "./common/border-radius.svelte";
 	import BorderStyle from "./common/border-style.svelte";
 	import BorderColor from "./common/border-color.svelte";
+    import EffectsDropShadow from "./common/effects-drop-shadow.svelte";
+	import EffectsShadowColor from "./common/effects-shadow-color.svelte";
+	import EffectsOpacity from "./common/effects-opacity.svelte";
+	import EffectsMixBlendMode from "./common/effects-mix-blend-mode.svelte";
+	import EffectsBackgroundBlendMode from "./common/effects-background-blend-mode.svelte";
 
     let globalComponentCollection = $globalComponentCollectionStore;
     $: globalComponentCollectionStore.set(globalComponentCollection);
@@ -142,6 +147,10 @@
         whitespace = "";
         wordbreak = "";
         borderStyle = "";
+        effectsDropShadow = "";
+        effectsOpacity = "";
+        effectsMixBlendMode = "";
+        effectsBackgroundBlendMode = "";
     }
 
 
@@ -385,6 +394,26 @@
      */
     let borderStyle;
 
+    /**
+     * @type string
+     */
+     let effectsDropShadow;
+
+    /**
+     * @type string
+     */
+    let effectsOpacity;
+
+    /**
+     * @type string
+     */
+    let effectsMixBlendMode;
+
+    /**
+     * @type string
+     */
+    let effectsBackgroundBlendMode;
+
     
     /**
      * @typedef {Object} SelectOptions
@@ -460,6 +489,11 @@
     let collapseDesignBorder = false;
     function toggleDesignBorder(){
         collapseDesignBorder = !collapseDesignBorder;
+    }
+
+    let collapseDesignEffects = false;
+    function toggleDesignEffects(){
+        collapseDesignEffects = !collapseDesignEffects;
     }
 
 </script>
@@ -713,6 +747,39 @@
 
     <div class="widgetPanelDivider"></div>
 
+    <button class="collapseButton" on:click={toggleDesignEffects}>
+        <span class="collapseHeader">EFFECTS</span>
+        {#if collapseDesignEffects}
+        <i class="bi bi-dash"></i>
+        {:else}
+        <i class="bi bi-plus"></i>
+        {/if}
+    </button>
+
+    {#if collapseDesignEffects}
+    <div class="w-full" in:slide={{ duration: 400 }} out:slide={{ duration: 100 }} >
+
+        <EffectsDropShadow bind:property={effectsDropShadow} bind:loaded bind:classInput elementDataLoaded={elementDataLoaded} on:updateClass={updateClass} />
+
+        <div class="widgetPanelDivider"></div>
+
+        <EffectsShadowColor bind:loaded bind:classInput elementDataLoaded={elementDataLoaded} on:updateClass={updateClass} />
+
+        <div class="widgetPanelDivider"></div>
+
+        <EffectsOpacity bind:property={effectsOpacity} bind:loaded bind:classInput elementDataLoaded={elementDataLoaded} on:updateClass={updateClass} />
+
+        <div class="widgetPanelDivider"></div>
+
+        <EffectsMixBlendMode bind:property={effectsMixBlendMode} bind:loaded bind:classInput elementDataLoaded={elementDataLoaded} on:updateClass={updateClass} />
+
+        <div class="widgetPanelDivider"></div>
+
+        <EffectsBackgroundBlendMode bind:property={effectsBackgroundBlendMode} bind:loaded bind:classInput elementDataLoaded={elementDataLoaded} on:updateClass={updateClass} />
+    </div>
+    {/if}
+
+    <div class="widgetPanelDivider"></div>
 
     <button class="collapseButton" on:click={toggleDesignClass}>
         <span class="collapseHeader">CLASS</span>